Remove dust emission at the domain edge
At the edge, the boundary data is transported into the domain. The grid cells
touching this edge are sometimes strongly influenced by the boundary. This
can cause problems with the dust emission. Therefore the emission at the
domain edge is not trustable.
I created an extra space with the size nboundlines at the edge where the
dust emission is suppressed.