The z0 dilemma
z0 is a complicated variable since it exists on multiple different scales
(synoptic, aeolian, local). See for instance Menut et al. (2013) JGR.

For the roughness correction (drag partitioning) in the Marticorena and
Bergametti scheme we need the aeolian z0. And the most consistent way would
be the use this z0 for the ustar calculation as well. However, the aeolian
z0 is really small which leads to very small ustar and an underestimation of
the dustflux. But in contrast, using the synoptic z0 form Cosmo does not
work for desert since it assumes relatively hight values (5cm) which
overestimates ustar and the dustflux.

As a workaround we use for now z0 = 1cm for the ustar calculation (NOT for
the drag partitioning). This value was suggested by Darmenova et al. (2009)
JGR.

Nevertheless, this can only be an interim solution.