Correct soil bulk density
For the soil moisture scheme it is necessary to transform the volumetric
soil moisture into the gravimetric soil moisture. Here the soil bulk density
is needed. Up to now the particle density was used (2650 kg/m3), which is
wrong! The new value is 1590 kg/m3 which is the standardised soil bulk
density for agricultural soil according to Ruehlmann and Körschens (2009),
Soil Science Society of America.

A better solution might be to use data set of soil bulk density, e.g. from
SoilGrids.