Merge soil and vegetation roughness
The model has two schemes for the drag partitioning (feff) caused by
1) the soil roughness and
2) the vegetation

On cropland may both appear depending the field is vegetated or not. So both
drag partitions must be merged somehow. The first approach was to multiply
them, but this seems incorrect since the soil roughness becomes irrelevant
once the vegetation is grown a bit. The new approach uses only the dominant
one with a min function.

feff = MIN(feff_z0, feff_veg)